Added auto-generation of HT Slots in HT
Now the Hash Table Manager would look through the
list of objects/triggers in the level and find all
the HexItems that are placed in the level. For each
HexItem, one HTSlot with the corresponding question
as the label would be generated.

Fixed the scaling of these auto generated HT Slots
by using transform.SetParent(someTransform, false)
instead of transform.parent = someTransform. This
should be done when setting the parent of any UI thing.

The generated HT Slots also contains data of which
is the correct HexItem should be in its slot,
but the checking of whether the item is correct
has not been done yet.