Added blank panel to HT ScrollRect, rm extra slots
Removed the extra HT slots that were used for testing.
Added a panel that is the size of a HTSlot which would
get added to the HTSlotsContainer after all the HTSlots
are added to extend its ScrollRect. This is to ensure
that no HTSlot would always be blocked by the inventory.

Updated the name of the HTSlotsList in HTManager to
reflect that only slots that are preadded to the HT
should be in that list since the rest are auto generated.