MovableEntityMover now constantly checks for tiles
Now Entities that are Movable will constantly check if
it has fully entered a tile for the first time. If yes then
it would trigger the onEntityEnterTileFully function in the
MapController that currently does not do anything.

This system is supposed to allow for implementation of interactions
between entities and tiles when they get entered. (eg: Box and water
tiles)