Add working Level Timer
Added a working timer that displays the time elapsed in the level in
MM:SS format. The timer pauses when the level is paused.

If the number of minutes elapsed or the number of seconds elapsed is
only 1 digit long, an additional 0 will be added to the front of the
number so that the timer does not look weird.

Added some untested functionality for saving th ebest time taken for
each level as a high score.