Update Pause Menu
Changed the Pause Menu background to a better looking translucent one so
that the player can continue looking at the game view even if the game
is paused.

Changed the Resume button to use a better image as it's background.
Changed font sizes and fonts for the Resume Button and the Paused text.