Fixed Best Time saving logic
The best time now only saves in 2 situations. The first one is when the
player collects more bonus coins compared to a previous run.
The second situation happens if the player collected the same amount
of bonus coins as the previous best and also if the time taken for the
current run is better.