Level Exit check when level begins, create TLCTT1
In case there are no HT Slots in some levels, the LMS now does a
check of whether the level exit should be unlocked at the start of the
level. The check is usually triggered by selection of items in the
inventory or Hash Table, since there are no items, this check is needed.